XMLHttpRequest/open-user-password-non-same-origin.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() - user/pass argument and non same-origin URL</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var m = "GET",
          u = "http://test2.w3.org/",
          a = false
      function request(opencall) {
        test(function() {
          var client = new XMLHttpRequest()
          assert_throws("INVALID_ACCESS_ERR", opencall(client))
        })
      }
      request(function(x) { return function() { x.open(m, u, a, "x") } })
      request(function(x) { return function() { x.open(m, u, a, "x", "x") } })
    </script>
  </body>
</html>
